+++ components/quickbooks_sandbox/quickbooks_sandbox.app.mjs
@@ -1,6 +1,7 @@
-import common from "../quickbooks/quickbooks.app.mjs";
+import common from "@pipedream/quickbooks";
 
 export default {
+  ...common,
   type: "app",
   app: "quickbooks_sandbox",
   methods: {